update to MITObim 1.6
-bug fix
--pairs option was not discriminating between headers and quality lines
starting with @. Lead to warnings.
Code contribution from A. Fischer (Max-Planck-Institute of Molecular
Plant Physiology, Potsdam, Germany) is much appreciated.
-bug fix
Subroutine that checks for length of supplied reference and splits it
in case it is longer than 29800bp (upper limit for mirabait) yielded
wrong results in some cases.
-bug fix
autostop in case of stationary read number went wrong if the assembly
contained several contigs
-new option (--454)
454 data is enabled.
-new option (--kbait)
Enables the user to modify the baiting stringency via the baiting kmer
Code contribution from M. Kweskin & E. Kayal (National Museum of Natural
History
Smithsonian Institution, Washington, DC, USA) is much appreciated.
-new option (--clean)
automatically removes every but the last two iteration directories.
useful if many iterations need to be run and diskspace is limited
-new option (--mirapath)
give the path to the mira binaries on the command line. Useful if for
some reason the particular mira version you want to use with MITObim
is not in your path.